+++ src/views/payment-method-view.js
@@ -3,7 +3,9 @@
 var BaseView = require('./base-view');
 var classlist = require('../lib/classlist');
 var constants = require('../constants');
-var paymentMethodHTML = require('../html/payment-method.html');
+var fs = require('fs');
+
+var paymentMethodHTML = fs.readFileSync(__dirname + '/../html/payment-method.html', 'utf8');
 
 function PaymentMethodView() {
   BaseView.apply(this, arguments);