+++ CHANGELOG.md
@@ -5,6 +5,7 @@ unreleased
 ----------
 - Use version 3.8.0 of braintree-web
 - Doesn't show Card payment option for merchants without cards enabled
+- Publish to npm
 
 1.0.0-beta.4
 ------------