+++ CHANGELOG.md
@@ -10,6 +10,7 @@ unreleased
 - Animate choosing a different way to pay
 - Switch PayPal component for PayPal Checkout component
 - Publish to npm
+- Fix #85 where the Drop-in would overflow in a small container
 
 1.0.0-beta.4
 ------------