+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 CHANGELOG
 =========
 
-unreleased
+1.0.0-beta.5
 ----------
 - Use version 3.10.0 of braintree-web
 - Doesn't show Card payment option for merchants without cards enabled